% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensate_all.R
\name{compensate_all}
\alias{compensate_all}
\title{Temperature and salinity compensation}
\usage{
compensate_all(
  base_temp = 25,
  temperature,
  dissolved_oxygen,
  conductivity,
  sal_comp = F
)
}
\arguments{
\item{base_temp}{The temperature (˚C) that dissolved oxygen probe measures against. Default 25˚C}

\item{temperature}{A vector or 1-D column of temperature measurements that coincide with conductivity measurements}

\item{dissolved_oxygen}{A vector or 1-D column of dissolved measurements that coincide with temperature measurements}

\item{conductivity}{A vector or 1-D column of conductivity measurements that coincide with temperature measurements.}

\item{sal_comp}{Logical denoting whether or not to do salinity compensation. Leave as FALSE unless salinity goes above 2}
}
\value{
Dissolved oxygen values adjusted for temperature, specific conductance values, and salinity values (optional)
}
\description{
\code{compensate_all} accepts dissolved oxygen,temperature, and
conductivity measurements and produces temperature compensated
dissolved oxygen, specific conductance, and salinity values (optional).
}
