% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO.R
\name{DO_adjust}
\alias{DO_adjust}
\title{Adjust dissolved oxygen}
\usage{
DO_adjust(
  base_temp = 25,
  temperature,
  dissolved_oxygen,
  salinity = rep(0, length(temperature))
)
}
\arguments{
\item{base_temp}{The temperature (˚C) that dissolved oxygen probe measures against. Default 25˚C}

\item{temperature}{A vector or 1-D column of temperature measurements that coincide with conductivity measurements}

\item{dissolved_oxygen}{A vector or 1-D column of dissolved measurements that coincide with temperature measurements}

\item{salinity}{A vector or 1-D column of salinity measurements that coincide with temperature measurements. Can be calculated
from conductivity and temperature using \code{\link{salinity}}}
}
\value{
Dissolved oxygen values adjusted for temperature
}
\description{
\code{DO_adjust} accepts dissolved oxygen and temperature measurements
and adjusts dissolved oxygen based on solubility of oxygen at
different temperatures assuming constant pressure. Uses USGS methods (citing Garcia and Gordon, 1992)
found here: \url{https://water.usgs.gov/admin/memo/QW/qw11.03.pdf}
}
