% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity.R
\name{salinity}
\alias{salinity}
\title{Calculate salinity using specific conductivity}
\usage{
salinity(specific_conductance)
}
\arguments{
\item{specific_conductance}{a vector or 1-D column of temperature measurements that coincide with conductivity measurements}
}
\value{
Practical salinity units (PSU)
}
\description{
\code{salinity} Accepts specific conductance values (uS/cm) and
converts them to practical salinity units (PSU, dimensionless). Equations from Schemel 2001 (as described here: https://pubs.usgs.gov/tm/2006/tm1D3/pdf/TM1D3.pdf)
}
