% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_conductance.R
\name{sp_conductance}
\alias{sp_conductance}
\title{Calculate specific conductance}
\usage{
sp_conductance(temperature, conductivity)
}
\arguments{
\item{temperature}{A vector or 1-D column of temperature measurements that coincide with conductivity measurements}

\item{conductivity}{A vector or 1-D column of conductivity measurements that coincide with temperature measurements}
}
\value{
Specific conductance values adjusted to 25˚C
}
\description{
\code{sp_conductance} accepts conductivity and temperature measurements
and converts them to values of specific conductance at 25˚C
}
